Convert system test to integration test
